Module documentation::UTF8

Re-exports

pub use crate::_Wrappers_Compile::Result;

Macros

int

seq

Structs

DafnyCharUTF16

DafnyInt

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

_default

Enums

Sequence

Traits

Default

A trait for giving a type a useful default value.

Functions

__init_ValidUTF8Bytes

integer_range

Type Aliases

ValidUTF8Bytes

nat

Derive Macros

Default

Derive macro generating an impl of the trait Default.