Module documentation::_Wrappers_Compile

Re-exports

pub use crate::_Wrappers_Compile::Option::None;

pub use crate::_Wrappers_Compile::Outcome::Fail;

pub use crate::_Wrappers_Compile::Result::Failure;

pub use crate::_Wrappers_Compile::Result::Success;

Structs

DafnyCharUTF16

Formatter

Configuration for formatting.

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

_default

Enums

Option

Outcome

Result

Sequence

Traits

AsRef

Used to do a cheap reference-to-reference conversion.

DafnyPrint

DafnyType

Debug

? formatting.

Default

A trait for giving a type a useful default value.

Eq

Trait for comparisons corresponding to equivalence relations.

Hash

A hashable type.

Hasher

A trait for hashing an arbitrary stream of bytes.

Functions

string_utf16_of

Derive Macros

Debug

Derive macro generating an impl of the trait Debug.

Default

Derive macro generating an impl of the trait Default.

Eq

Derive macro generating an impl of the trait Eq.

Hash

Derive macro generating an impl of the trait Hash.